backend/pyproject.toml
[project]
name = "talk2doc-backend"
version = "0.1.0"
description = "Talk2Doc Python backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain>=0.3.19",
    "langchain-openai>=0.2.14",
    "langchain-postgres>=0.0.13",
    "langchain-text-splitters>=0.3.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "pypdf>=5.1.0",
    "docx2txt>=0.8",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.3",
    "alembic>=1.18.3",
    "pydantic-settings>=2.12.0",
    "pgvector>=0.3.6",
    "pymupdf>=1.26.7",
    "uvicorn[standard]>=0.40.0",
    "openai>=2.16.0",
    "langchain-experimental>=0.4.1",
    "networkx>=3.6.1",
    "langgraph>=1.0.7",
    "pandas>=3.0.0",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.2",
]

[project.scripts]
dev = "uvicorn main:app --reload --port 8000"
